+++ converter/rst/extrtoolembed.py
@@ -1,0 +1,28 @@
+import re
+from collections import namedtuple
+
+AssessmentData = namedtuple('AssessmentData', ['id', 'name', 'type', 'points', 'ex_data'])
+
+
+class ExtrToolEmbed(object):
+    def __init__(self, source_string, exercises):
+        self.str = source_string
+        self._exercises = exercises
+        self._assessments = list()
+        self._extrtoolembed_re = re.compile(
+            r"""^$\n^.*?\n-+\n\n?\.\. extrtoolembed:: '(?P<name>.*?)'\n( *:.*?: .*?\n)?(?=\S|$)""", flags=re.MULTILINE)
+
+    def _extrtoolembed(self, matchobj):
+        name = matchobj.group('name').lower()
+        ex_data = self._exercises.get(name.lower(), {})
+        assessment_id = f'test-{name.lower()}'
+        if not ex_data:
+            return ''
+        assessment = AssessmentData(assessment_id, name, 'test', 20, ex_data)
+        self._assessments.append(assessment)
+        return ''
+
+    def convert(self):
+        output = self.str
+        output = self._extrtoolembed_re.sub(self._extrtoolembed, output)
+        return output, self._assessments
